Add get_attachment_ratio_image template function.
This function allows you to wrap a lazy loaded image in a container with an intrinsic ratio applied via an inline style with padding-bottom.

There’s currently no way to filter the final html output of the wp_get_attachment_image function so I’ve added this template function until a html filter is added.

@link https://core.trac.wordpress.org/attachment/ticket/27399/43543.patch